Runner 3D to Runner AR
The 3D Runner is being adapted to AR. For now, the input has been refactored to include the most current Inputs system so that “OnScreen” scripts can be used for UI elements such as buttons and joysticks (of which there is already one). The logic has also been modified so that the chunks move toward the player, rather than the player constantly moving forward.